Dala 4697 allow range of pages (#676)
This changes the format of a page for the Frameworks data model.
Previously, the format was an integer describing a single page of a document.
Now the format is changed to a string which still allows single pages, e.g. "5", "89", "41" etc, but also ranges "14-15", "1-100" etc. 
We have also fixed the fact that the AssuranceDataPoints were not validated in the EU Taxonomy for non-financial companies.

---------

Co-authored-by: Tayfun Dertsiz <Tayfun.Dertsiz@d-fine.com>
Co-authored-by: d93476 <jakob.kraus@d-fine.com>
Co-authored-by: Tim Breitenfelder <tim.breitenfelder@d-fine.com>
Co-authored-by: Laurin Strelow <1918073+LaurinStrelow@users.noreply.github.com>